+++ widgets/src/layout/expanded.rs
@@ -15,18 +15,21 @@ impl<'c> ComposeChild<'c> for Expanded {
   type Child = Widget<'c>;
   #[inline]
   fn compose_child(this: impl StateWriter<Value = Self>, child: Self::Child) -> Widget<'c> {
-    fn_widget! {
-      @ConstrainedBox {
-        clamp: BoxClamp {
-          min: Size::new(0., 0.),
-          max: Size::new(f32::INFINITY, f32::INFINITY)
-        },
-        @{ child }
-      }
+    let data: Box<dyn Query> = match this.try_into_value() {
+      Ok(data) => Box::new(Queryable(data)),
+      Err(data) => Box::new(data),
+    };
+    Provider::new(data)
+      .with_child(fn_widget! {
+        @ConstrainedBox {
+          clamp: BoxClamp {
+            min: Size::new(0., 0.),
+            max: Size::new(f32::INFINITY, f32::INFINITY)
+          },
+          @{ child }
+        }
+      })
       .into_widget()
-      .try_unwrap_state_and_attach(this)
-    }
-    .into_widget()
   }
 }